add json_ prefix to strbuffer_, hashtable_, and utf8_ methods
Summary: C doesn't have namespaces, so libjansson.a should to use symbol
names that are unlikely to be used elsewhere.  After this diff nm shows
that all symbols start with "json", which is probably sufficient.  None
of the renamed functions are part of the public API, they are only used
internally.